GUI for choosing sound files completed.
Currently uses the 'white' and 'black' buttons to play sound.
If another sound button is pushed, and a sound is currently playing,
it is ignored. Choosing a new sound file using the dialogs will
change the sound files that are played by the dalek.

Also added a line to blank the screen before we draw, so that
we don't get ghost dialog boxes floating around. Also improves the
text display sharpness, which was just beginning to annoy me.
 Please enter the commit message for your changes. Lines starting