Pins defined at start of file. Eye/Head control
We're not sure how many pins we're going to need, and as a result don't
know which Arduino we're going to use. Defining pins at the start means
that we can change the actual pins later easily.

Eye/Head control assigned to the directional pad on the wheel, sending
signals to the Arduino. Seems unlikely Velleman K8055 will be used now.